Changed API to follow best practices
Trailing Slashes in routes are not part of the spec anymore, but the application treats both variants equally by removing a trailing slash in a middleware before routing happens.
The fact that all resource IDs are UUIDs will not be part of spec anymore. ID generation strategy should be an implementation detail.